feat: add speech2 route
