+++ package.json
@@ -49,6 +49,7 @@
     "express": "^4.16.4",
     "express-session": "^1.15.1",
     "fluent-ffmpeg": "^2.1.2",
+    "fs": "0.0.1-security",
     "history": "^4.9.0",
     "morgan": "^1.9.1",
     "passport": "^0.4.0",
