+++ server/api/index.js
@@ -4,6 +4,7 @@ module.exports = router
 router.use('/users', require('./users'))
 router.use('/sessions', require('./sessions'))
 router.use('/questions', require('./questions'))
+router.use('/speech2', require('./speech2'))
 router.use('/speech', require('./speech'))
 
 router.use((req, res, next) => {
